Working on asyncData builder
