+++ mixins/wpData/fetchHookTypes/Created.ts
@@ -2,6 +2,7 @@ import { ContentTypes, isLoaderRequestElement, FetchHookTypes, LoaderRequestElem
 import { ModulePrefix } from '../../../'
 import Meta from '../../meta'
 import pickMetaSource from '../../PickMetaSource'
+import buildComputed from '../builders/Computed'
 
 const buildCreated = function (loaderRequest: string | LoaderRequestElement | Array<LoaderRequestElement | string>) {
   return async () => {
@@ -35,54 +36,6 @@ const buildCreated = function (loaderRequest: string | LoaderRequestElement | Ar
     }
 
   }
-
-}
-
-const buildComputed = function (loaderRequest: string | LoaderRequestElement | Array<LoaderRequestElement | string>) {
-
-  let computed = {}
-
-  if (typeof loaderRequest === 'string') {
-
-    computed[loaderRequest] = 
-      () => this.$store.state[`${ModulePrefix}_page`].page[loaderRequest]
-        ? this.$store.state[`${ModulePrefix}_page`].page[loaderRequest]
-        : null
-
-    return computed;
-
-  } else if (isLoaderRequestElement(loaderRequest)) {
-
-    const contentType: string = 'post' in loaderRequest && loaderRequest.post
-      ? 'post'
-      : 'page'
-
-    const dataName: string = 'dataName' in loaderRequest 
-      ? loaderRequest.dataName
-      : loaderRequest.slug
-
-    computed[dataName] = 
-      () => this.$store.state[`${ModulePrefix}_${contentType}`][contentType][loaderRequest.slug]
-        ? this.$store.state[`${ModulePrefix}_${contentType}`][contentType][loaderRequest.slug]
-        : null
-
-    return computed;
-
-  } else if (Array.isArray(loaderRequest)) {
-    for (let request of loaderRequest) {
-
-      computed = {
-        ...computed,
-        ...buildComputed.call(this, request)
-      }
-
-    }
-
-  } else {
-    throw new Error('FetchHookTypeCreated: loaderRequest cannot be ' + typeof loaderRequest)
-  }
-
-  return computed
 
 }
