+++ mixins/wpData/fetchHookTypes/VoidAsyncData.ts
@@ -1,9 +1,21 @@
-export default () => {
+import { ContentTypes, isLoaderRequestElement, FetchHookTypes, LoaderRequestElement } from '../../../types'
+import { ModulePrefix } from '../../../'
+import Meta from '../../meta'
+import pickMetaSource from '../../PickMetaSource'
+import buildComputed from '../builders/Computed'
+import buildAsyncData from '../builders/AsyncData'
+
+export default function (loaderRequest: string | LoaderRequestElement | Array<LoaderRequestElement | string>) {
+  const asyncData = buildAsyncData.call(this, loaderRequest, FetchHookTypes.VoidAsyncData)
+  const computed = buildComputed.call(this, loaderRequest)
+  const meta = pickMetaSource.call(this, loaderRequest)
 
   return {
-    asyncData () {
-      
-    }
+    asyncData,
+    computed,
+    mixins: [
+      Meta(ContentTypes.Page, meta)
+    ]
   }
 
 }
