+++ types/index.ts
@@ -21,3 +21,7 @@ export interface LoaderRequestElement {
 export function isLoaderRequestElement(object: any): object is LoaderRequestElement {
   return typeof object === 'object' && 'slug' in object
 }
+
+export interface LoaderRequestElementWithValue extends LoaderRequestElement{
+  value: any
+}
